% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbListForeignKeys.R
\name{dbListForeignKeys}
\alias{dbListForeignKeys}
\alias{dbListForeignKeys,DBIConnection,Id-method}
\alias{dbListForeignKeys,SQLiteConnection,Id-method}
\title{List Foreign Keys}
\usage{
dbListForeignKeys(conn, id, ...)

\S4method{dbListForeignKeys}{DBIConnection,Id}(conn, id, ...)

\S4method{dbListForeignKeys}{SQLiteConnection,Id}(conn, id, ...)
}
\arguments{
\item{conn}{a \code{DBIConnection}.}

\item{id}{an \code{\link[DBI]{Id}} specifying a database table.}

\item{\dots}{additional arguements.}
}
\description{
List the foreign keys of a database table.
}
\section{Value}{
 a compact representation of the foreign keys.
}

