% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbListPrimaryKeys.R
\name{dbListPrimaryKeys}
\alias{dbListPrimaryKeys}
\alias{dbListPrimaryKeys,DBIConnection,Id-method}
\alias{dbListPrimaryKeys,SQLiteConnection,Id-method}
\title{List Primary Keys}
\usage{
dbListPrimaryKeys(conn, id, ...)

\S4method{dbListPrimaryKeys}{DBIConnection,Id}(conn, id, ...)

\S4method{dbListPrimaryKeys}{SQLiteConnection,Id}(conn, id, ...)
}
\arguments{
\item{conn}{a \code{DBIConnection}.}

\item{id}{an \code{\link[DBI]{Id}} specifying a database table.}

\item{\dots}{additional arguements.}
}
\description{
List the primary keys of a database table.
}
\section{Value}{
 a character vector containing the names of the primary
                key columns.
}

