% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbListSchema.R
\name{dbListSchema}
\alias{dbListSchema}
\alias{dbListSchema,DBIConnection-method}
\alias{dbListSchema,DBIConnection,Id-method}
\title{List Database Schema}
\usage{
dbListSchema(conn, prefix = NULL, ...)

\S4method{dbListSchema}{DBIConnection}(conn, prefix = NULL, ...)
}
\arguments{
\item{conn}{a \code{DBIConnection}.}

\item{prefix}{an \code{\link[DBI]{Id}} specifying fully qualified path
in the database's namespace, or NULL.}

\item{\dots}{additional arguements.}
}
\description{
List a database schema by table including column names.
}
\section{Value}{
 a compact representation of the database schema.
}

