% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_by_id}
\alias{split_by_id}
\title{Compact Storage for DBI Info}
\usage{
split_by_id(x, idcols)
}
\arguments{
\item{x}{a \code{\link{data.frame}} where one or more of the columns
identifies a \emph{table}.}

\item{idcols}{a character vector specifying the subset of the columns
of \code{x} that will be passed to \code{\link[DBI]{Id}}.}
}
\description{
Store database \emph{Info} in a compact data structure
             indexed by an \code{\link[DBI]{Id}}. This function is
             provided to simplify creating methods for DBI-extending
             packages.
}
